requirements.txt
AgentLogin==3.0.0
asgiref==3.6.0
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.15
django-haystack==3.2.1
django-redis==5.2.0
elasticsearch==2.4.1
idna==3.4
itsdangerous==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
Pillow==10.4.0
PyMySQL==1.0.3
pytz==2023.3
redis==4.5.4
requests==2.30.0
sqlparse==0.4.4
urllib3==1.26.15
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
django-filter==23.2
